Refactor categorization algorithm
Fulfill smallest groups first. So we can't have files fulfilling the
yearly group before we're done with the monthly group, for example.

In practice, this means something like -keep-monthly 12 -keep-yearly 2
would keep the most recent 12 months and an additional 2 years.
Previously this worked a little differently and it was more confusing to
know exactly how many files you could end up with. This will be more
precise.

We've also added some more detail and colors to the stdout to make it
easier to see what's happening.